agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain-core",
    "openai",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "pydantic>=2.9.0",
    "boto3",
    "orjson",
    "sse-starlette",
    "supabase",
    "nest-asyncio",
    "moviepy",
    "pyyaml",
    "pydub",
    "pyjwt>=2.8.0",
    "pillow>=12.1.1",
    "python-pptx>=0.6.23",
    "pdfplumber>=0.11.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.2.0",
]
